+++ models/images.go
@@ -18,7 +18,7 @@ import (
 //Images Model
 type Images struct {
 	ID        int         `orm:"column(id);pk" json:"id"`
-	Priority  int8        `orm:"column(priority)" json:"priority,omitempty"`
+	Priority  int         `orm:"column(priority)" json:"priority,omitempty"`
 	Name      string      `orm:"column(name);size(255)" json:"name,omitempty" valid:"Required"`
 	Slug      string      `orm:"column(slug);size(255)" json:"slug,omitempty" valid:"AlphaDash"`
 	UUID      string      `orm:"column(uuid);size(255)" json:"uuid,omitempty" valid:"Required"`
@@ -240,6 +240,7 @@ func DeleteImages(id int, trash bool) (err error) {
 	}
 
 	if trash {
+
 		_, err = o.Delete(&v)
 
 		if err == nil {
